# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MyBlog API
  version: 1.0.0
  description: Vue 前端已验证的数据契约。评论接口由 Artalk 独立提供。
servers:
  - url: /api
...
paths:
...
  /admin/uploads:
    get:
      tags: [Uploads]
      summary: 获取媒体库
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { type: string, enum: [active, trashed], default: active }
        - name: usage
          in: query
          schema: { type: string, enum: [all, used, unused], default: all }
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
      responses:
        '200':
          description: 媒体分页
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadPageResponse' }
    post:
      tags: [Uploads]
      summary: 上传图片或允许的附件
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadResponse' }
        '413': { $ref: '#/components/responses/BadRequest' }
        '415': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: blog_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
...
  responses:
    BadRequest:
      description: 请求参数或字段错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 1 }
...
    Upload:
      type: object
      required: [id, url, filename, contentType, size, width, height, status, usageCount, references, createdAt, trashedAt]
      properties:
        id: { type: integer, format: int64 }
        url: { type: string }
        filename: { type: string }
        contentType: { type: string }
        size: { type: integer, format: int64, minimum: 1 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        status: { type: string, enum: [active, trashed] }
        usageCount: { type: integer, minimum: 0 }
        references:
          type: array
          items: { $ref: '#/components/schemas/UploadReference' }
        createdAt: { type: string, format: date-time }
        trashedAt: { type: [string, 'null'], format: date-time }
    UploadReference:
      type: object
      required: [resourceType, resourceId, field, label]
      properties:
        resourceType:
          type: string
          enum: [site_avatar, site_banner, site_about, post_cover, post_content]
        resourceId: { type: [integer, 'null'], format: int64 }
        field: { type: string }
        label: { type: string }
    UploadPage:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Upload' }
        pagination: { $ref: '#/components/schemas/Pagination' }
    UploadPageResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/UploadPage' }
    UploadResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Upload' }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
        requestId: { type: string }